Revert "Use nfs mounts (#14)"
Unless abhi has a reason to keep this as is.  Right now, this isn't the
pattern we're following in other projects, and as it is it also embeds
local minikube configuration details that may vary from host to host in
its helm chart (specifically, the minikube ip and the fact that nfs is
configured).

This reverts commit b73facc031613d5b48cdd3f4c47be425f8dbdef5.